requirements.txt
discord~=1.0.0
python-dotenv~=0.14.0
tabulate~=0.8.7